Remove unuseful hooks
