+++ packages/react-topology/src/utils/index.ts
@@ -10,9 +10,5 @@ export { default as useCombineRefs } from './useCombineRefs';
 export { default as useHover } from './useHover';
 export * from './useCustomNodeShape';
 export * from './useSize';
-export * from './useStatus';
-export * from './useLabel';
-export * from './useBadge';
 export * from './useCollapsibleGroup';
 export * from './createAggregateEdges';
-export * from './useTerminals';
